requirements.txt
# Core Dependencies
langgraph==1.0.2
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-google-genai>=0.0.6
langchain-anthropic>=0.0.2
langchain-core>=0.1.16

# Web Framework
streamlit==1.29.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings>=2.1.0
requests==2.31.0
tenacity==8.2.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Logging & Monitoring
loguru==0.7.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

